+++ solution/RenderEngine/include/IEditorWindow.h
@@ -9,7 +9,7 @@ class IEditorWindow
 {
 public:
 	virtual void onInit() {}
-	virtual void onUpdate(uint32_t elapsedTime) {}
+	virtual void onUpdate(float elapsedTime) {}
 	virtual void onEditorGUI() = 0;
 	virtual void onRelease() {}
 };
